# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vending machine management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stock/movements:
    get:
      operationId: listStockMovements
      tags: [stock]
      summary: List stock movements
      parameters:
        - name: productId
          in: query
          schema:
            type: ["integer", "null"]
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
    post:
      operationId: createStockMovement
      tags: [stock]
      summary: Add or remove stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockMovementInput"
      responses:
        "201":
          description: Movement recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovement"

  # ── Machine Loads (Sales) ─────────────────────────────────────────────────
...
components:
  schemas:
...
    StockMovement:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        type:
          type: string
          enum: [in, out]
        reason:
          type: string
          enum: [purchase, load, official_entry, official_exit, expired, defect, other]
        quantity:
          type: number
        clientMachineId:
          type: ["integer", "null"]
        clientMachineName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        operatorId:
          type: ["integer", "null"]
        operatorName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - productId
        - productName
        - type
        - reason
        - quantity
        - createdAt
    StockMovementInput:
      type: object
      properties:
        productId:
          type: integer
        type:
          type: string
          enum: [in, out]
        reason:
          type: string
          enum: [purchase, load, official_entry, official_exit, expired, defect, other]
        quantity:
          type: number
        clientMachineId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        operatorId:
          type: ["integer", "null"]
      required:
        - productId
        - type
        - reason
        - quantity

    # ── Machine Load ─────────────────────────────────────────────────────────
...