# lib/api-zod/src/generated/types/stockMovement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vending machine management API
 * OpenAPI spec version: 0.1.0
 */
import type { StockMovementReason } from './stockMovementReason';
import type { StockMovementType } from './stockMovementType';

export interface StockMovement {
  id: number;
  productId: number;
  productName: string;
  type: StockMovementType;
  reason: StockMovementReason;
  quantity: number;
  /** @nullable */
  clientMachineId?: number | null;
  /** @nullable */
  clientMachineName?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  operatorId?: number | null;
  /** @nullable */
  operatorName?: string | null;
  createdAt: string;
}